feat: Issue #319対応 - 持続可能なアーキテクチャへの大規模リファクタリング
## 概要
技術的負債を解消し、定期的なリファクタリングを不要にする持続可能なアーキテクチャを確立

## 主な変更内容

### アーキテクチャ原則の策定
- docs/dev/ARCHITECTURE_PRINCIPLES.md 作成
- 300行制限、単一責任原則、依存方向規則の明文化
- インターフェース駆動設計の採用

### 自動品質管理システムの導入
- dev/tools/architecture_validator.py - アーキテクチャ検証ツール
- dev/tools/check_file_size.py - ファイルサイズ監視
- .pre-commit-config.yaml - 自動品質チェック

### 大規模ファイルのモジュール分割（300行超→小規模モジュール）
1. error_reporting.py (498行→18行)
   - core/reporting/: 型定義、フォーマット、レポート、コンテキスト

2. commands/convert.py (375行→44行)
   - commands/convert/: コマンド制御、検証、処理、監視

3. config_manager.py (342行→14行)
   - core/config/: 型定義、検証、読込、管理

4. distribution_manager.py (371行→19行)
   - core/distribution/: 構造、変換、処理、管理

5. smart_cache.py (303行→15行)
   - core/caching/: 型定義、戦略、ストレージ、キャッシュ

## 成果
- サイズ制限違反: 5件→0件（100%解消）
- 平均削減率: 94%
- 後方互換性: 100%維持
- 持続可能な開発基盤の確立

Closes #319
