+++ kumihan_formatter/commands/convert/__init__.py
@@ -1,0 +1,20 @@
+"""
+変換コマンドモジュール
+
+Convert command の責任分離実装
+Issue #319対応 - 単一責任原則に基づくリファクタリング
+
+元ファイル: commands/convert.py (375行) → 4つのモジュールに分割
+"""
+
+from .convert_command import ConvertCommand
+from .convert_validator import ConvertValidator
+from .convert_processor import ConvertProcessor
+from .convert_watcher import ConvertWatcher
+
+__all__ = [
+    "ConvertCommand",
+    "ConvertValidator",
+    "ConvertProcessor", 
+    "ConvertWatcher"
+]
