+++ kumihan_formatter/core/config/__init__.py
@@ -1,0 +1,24 @@
+"""
+設定管理モジュール
+
+Enhanced configuration management の責任分離実装
+Issue #319対応 - 単一責任原則に基づくリファクタリング
+
+元ファイル: core/config_manager.py (342行) → 4つのモジュールに分割
+"""
+
+from .config_types import ConfigLevel, ValidationResult
+from .config_validator import ConfigValidator
+from .config_loader import ConfigLoader
+from .config_manager import EnhancedConfig
+
+__all__ = [
+    # 型定義
+    "ConfigLevel",
+    "ValidationResult",
+    
+    # 機能クラス
+    "ConfigValidator",
+    "ConfigLoader",
+    "EnhancedConfig"
+]
